refactor : recent API 모두 user 로 이동
* recent 관련 API 모두 USER 모듈로 이동
* recent_played 테이블을 따로 만든 것을 고려
