+++ server/src/playlist/playlist.module.ts
@@ -7,11 +7,10 @@ import { AuthModule } from 'src/auth/auth.module';
 import { Music_Playlist } from 'src/entity/music_playlist.entity';
 import { Music } from 'src/entity/music.entity';
 import { Logger } from 'winston';
-import { Recent_Played } from 'src/entity/recent_played.entity';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Playlist, Music_Playlist, Music, Recent_Played]),
+    TypeOrmModule.forFeature([Playlist, Music_Playlist, Music]),
     AuthModule,
   ],
   controllers: [PlaylistController],
