+++ server/src/user/user.controller.ts
@@ -9,6 +9,7 @@ import {
   Body,
   Query,
   Logger,
+  Put,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { HTTP_STATUS_CODE } from 'src/httpStatusCode.enum';
@@ -90,5 +91,23 @@ export class UserController {
   ): Promise<User[]> {
     this.logger.log(`GET /users/search - keyword=${keyword}`);
     return this.userService.getCertainKeywordNicknameUser(keyword);
+  }
+
+  @Put('recent-played')
+  @UseGuards(AuthGuard())
+  @HttpCode(HTTP_STATUS_CODE.SUCCESS)
+  async updateRecentPlayMusic(
+    @Req() req,
+    @Body('musicId') music_id: string,
+  ): Promise<{ recent_played_id: number }> {
+    this.logger.log(
+      `PUT /playlists/recent-played - nickname=${req.user.nickname}, music_id=${music_id}`,
+    );
+    const user_id: string = req.user.user_id;
+    const recent_played_id: number = await this.userService.updateRecentMusic(
+      music_id,
+      user_id,
+    );
+    return { recent_played_id };
   }
 }
