+++ server/src/user/user.module.ts
@@ -5,9 +5,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from 'src/entity/user.entity';
 import { AuthModule } from 'src/auth/auth.module';
 import { Logger } from 'winston';
+import { Recent_Played } from 'src/entity/recent_played.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User]), AuthModule],
+  imports: [TypeOrmModule.forFeature([User, Recent_Played]), AuthModule],
   controllers: [UserController],
   providers: [UserService, Logger],
 })
